requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
coreapi==2.3.3
coreschema==0.0.4
ddt==1.2.1
decorator==4.4.0
demjson==2.2.4
Django==3.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-filter-stubs==0.1.1
django-rest-swagger==2.2.0
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
drf-yasg==1.17.0
elasticsearch==7.8.0
et-xmlfile==1.0.1
Faker==4.0.3
filetype==1.0.5
gevent==20.5.0
greenlet==0.4.15
grequests==0.6.0
har2case==0.3.1
httpie==1.0.3
HttpRunner==1.5.8
idna==2.8
importlib-metadata==1.4.0
inflection==0.3.1
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonpath==0.82
jsonschema==3.2.0
kazoo==2.7.0
livetest==0.5
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
mypy==0.770
mypy-extensions==0.4.3
mysqlclient==1.4.4
numpy==1.18.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==19.2
pandas==1.0.1
parso==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==2.0.9
py==1.8.1
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.2
PyPubSub==3.3.0
pyrsistent==0.15.7
pytest==5.4.1
pytest-html==2.1.1
pytest-metadata==1.8.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
PyYAML==5.3
requests==2.22.0
requests-toolbelt==0.9.1
robotframework==3.1.2
robotframework-ride==1.7.4
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.1.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
selenium==3.141.0
sentry-sdk==0.13.5
simplejson==3.17.2
six==1.12.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.2
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.7
wxPython==4.0.4
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==2.0.0
